Extract world model query routing
